Convert bluetooth::shim::BtmStatus from 16 to 8 bits
Align with rest of stack.

Bug: 166280067
Test: gd/cert/run
Tag: #refactor
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 lines

+++ main/shim/btm.h
@@ -82,7 +82,7 @@ using HACK_ScoDisconnectCallback = std::function<void(uint16_t, uint8_t)>;
 namespace bluetooth {
 namespace shim {
 
-using BtmStatus = enum : uint16_t {
+using BtmStatus = enum : uint8_t {
   BTM_SUCCESS = 0,              /* Command succeeded                 */
   BTM_CMD_STARTED = 1,          /* Command started OK.               */
   BTM_BUSY = 2,                 /* Device busy with another command  */
